problem_1/geonode/requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
billiard==4.1.0
celery==5.3.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron_descriptor==1.4.0
Django==4.2.5
django-admin-rangefilter==0.12.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-timezone-field==6.0.1
h11==0.14.0
idna==3.6
kombu==5.3.2
outcome==1.3.0.post0
packaging==24.0
prompt-toolkit==3.0.39
pycparser==2.21
PySocks==1.7.1
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.1
redis==5.0.1
requests==2.31.0
selenium==4.18.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.4.4
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.10.0
tzdata==2023.3
urllib3==2.2.1
vine==5.0.0
wcwidth==0.2.6
webdriver-manager==4.0.1
wsproto==1.2.0
